% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{archiveFile}
\alias{archiveFile}
\title{Archive a file into the recordr archive directory}
\usage{
archiveFile(file)
}
\arguments{
\item{file}{The file to save in the archive}
}
\value{
The name of the archived file path relative to the recordr home directory
}
\description{
Archive a file into the recordr archive directory
}
\note{
This function is intended to run only during a record() session, i.e. the
recordr environment needs to be available.
}

