% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recordr.R
\name{changeHome}
\alias{changeHome}
\title{Change the recordr home directory}
\usage{
changeHome(recordr, currentDir, newDir = as.character(NA), copy, ...)
}
\arguments{
\item{recordr}{A recordr object}

\item{currentDir}{A character value specifying the current recordr home directory}

\item{newDir}{A character value, specifying the new recordr home directory}

\item{copy}{A logical value. A value of TRUE causes data to be copied from the old}

\item{...}{Additional arguments
directory to the new one. A default value is not set.}
}
\description{
Change the recordr home directory
}
