% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recordr.R
\name{coverageElement}
\alias{coverageElement}
\title{Create a coverage element}
\usage{
coverageElement(gc, tempc)
}
\arguments{
\item{gc}{An EML::geographicCoverage object}

\item{tempc}{A EML::temporalCoverage object}
}
\value{
An EML::Coverage object
}
\description{
Create a coverage element
}

