% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecMetadata.R
\docType{class}
\name{ExecMetadata-class}
\alias{ExecMetadata-class}
\title{A class representing a script execution with the run manager}
\description{
A class representing a script execution with the run manager
}
\section{Methods}{

\itemize{
 \item{\code{\link[=initialize-ExecMetadata]{initialize}}}{: Initialize an execution metadata object}
 \item{\code{\link{readExecMeta}}}{: Retrieve saved Execution metadata.}
 \item{\code{\link{writeExecMeta}}}{: Save a single execution metadat.}
 \item{\code{\link{updateExecMeta}}}{: Update saved execution metadata.}
}
}
\author{
slaughter
}
\seealso{
\code{\link{recordr}}{ package description.}
}

