% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recordr.R
\name{geoCoverage}
\alias{geoCoverage}
\title{Create a geographic coverage element from a description and bounding coordinates}
\usage{
geoCoverage(geoDescription, west, east, north, south)
}
\description{
Create a geographic coverage element from a description and bounding coordinates
}

