% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recordr.R
\name{getDBconnection}
\alias{getDBconnection}
\title{Get a database connection}
\usage{
getDBconnection(dbFile)
}
\arguments{
\item{dbFile}{the path to the recordr database file (default: ~/.recordr/recordr.sqlite)}
}
\description{
Get a database connection
}
