% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recordr.R
\docType{methods}
\name{getMetadata}
\alias{getMetadata}
\alias{getMetadata,Recordr-method}
\title{Retrieve the metadata object for a run}
\usage{
getMetadata(recordr, ...)

\S4method{getMetadata}{Recordr}(recordr, id = as.character(NA),
  seq = as.character(NA), as = as.character("text"))
}
\arguments{
\item{recordr}{a Recordr instance
seealso \code{\link[=Recordr-class]{Recordr}} { class description}}

\item{id}{The identifier for a run}

\item{seq}{The sequence number for a run}

\item{as}{Form to return the metadata as. Possible values are: "text", "parsed" (for parsed XML)}
}
\value{
A character vector containing the metadata
}
\description{
When a script or console session is recorded (see record() and startrecord()), 
a metadata object is created that describes the objects associated with the run, using the
Ecological Metadata Language \link{https://knb.ecoinformatics.org/#external//emlparser/docs/index.html}.
This metadata can be retrieved from the recordr cache for review or editing if desired. If the metadata
is updated, it can be re-inserted into the recordr cache using the \code{putMetadata} method.
}

