% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{getProvCapture}
\alias{getProvCapture}
\title{Return current state of provenance capture}
\usage{
getProvCapture(x)
}
\value{
enabled a logical indicating the state of provenance capture: TRUE=enabled, FALSE=disabled
}
\description{
Return current state of provenance capture
}

