% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecMetadata.R
\docType{methods}
\name{initialize,ExecMetadata-method}
\alias{initialize,ExecMetadata-method}
\alias{initialize-ExecMetadata}
\title{Initialize an execution metadata object}
\usage{
\S4method{initialize}{ExecMetadata}(.Object, executionId = as.character(NA),
  metadataId = as.character(NA), tag = as.character(NA),
  datapackageId = as.character(NA), user = as.character(NA),
  subject = as.character(NA), hostId = as.character(NA),
  startTime = as.character(NA), operatingSystem = as.character(NA),
  runtime = as.character(NA), moduleDependencies = as.character(NA),
  programName = as.character(NA), endTime = as.character(NA),
  errorMessage = as.character(NA), publishTime = as.character(NA),
  publishNodeId = as.character(NA), publishId = as.character(NA),
  console = FALSE, seq = as.integer(0))
}
\arguments{
\item{.Object}{The ExecMetada object}

\item{tag}{A character vector that describes this execution.}

\item{programName}{The name of the program that is being run.}
}
\description{
Initialize an execution metadata object
}
\seealso{
\code{\link[=ExecMetadata-class]{ExecMetadata}} { class description}
}

