% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileMetadata.R
\docType{methods}
\name{initialize,FileMetadata-method}
\alias{initialize,FileMetadata-method}
\alias{initialize-FileMetadata}
\title{Initialize a file metadata object.}
\usage{
\S4method{initialize}{FileMetadata}(.Object, file, fileId = as.character(NA),
  sha256 = as.character(NA), size = as.numeric(0),
  user = as.character(NA), createTime = as.character(NA),
  modifyTime = as.character(NA), executionId, access = as.character(NA),
  format = as.character(NA), archivedFilePath = as.character(NA))
}
\arguments{
\item{.Object}{a FileMetdata object}

\item{file}{The file to get metadata for}

\item{executionId}{the execution that this file belongs to}

\item{access}{the access that occurred for this file ("read", "write", "execute")}
}
\description{
Initialize a file metadata object.
}
\details{
This method is used internally by the recordr package.
}
\seealso{
\code{\link[=FileMetadata-class]{FileMetadata}} { class description}
}

