% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvRels.R
\docType{methods}
\name{initialize,ProvRels-method}
\alias{initialize,ProvRels-method}
\alias{initialize-ProvRels}
\title{Initialize a provenance relationship object.}
\usage{
\S4method{initialize}{ProvRels}(.Object, executionId = as.character(NA),
  subject = as.character(NA), predicate = as.character(NA),
  object = as.character(0), subjectType = as.character(NA),
  objectType = as.character(NA), dataTypeURI = as.character(NA))
}
\arguments{
\item{.Object}{a \code{"ProvRels"} object}

\item{executionId}{a \code{"character"}, the execution identifier for an execution}

\item{subject}{a \code{"character"}, the<BS>subject of the provenance relationship}

\item{predicate}{a \code{"character"}, the predicate of the provenance relationship}

\item{object}{a \code{"character"}, the object of the provenance relationship}

\item{subjectType}{a \code{"character"}, the RDF node type of the subject, i.e. "url", "blank"}

\item{objectType}{a \code{"character"}, the RDF node type of the object i.e. "url", "blank"}

\item{dataTypeURI}{a \code{"character"}, the RDF data type of the object, i.e. "xsd:string"}
}
\description{
Initialize a provenance relationship object.
}
\details{
This method is used internally by the recordr package.
}
\seealso{
\code{\link[=ProvRels-class]{ProvRels}} { class description}
}
