% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Recordr.R
\docType{methods}
\name{initialize,Recordr-method}
\alias{initialize,Recordr-method}
\alias{initialize-Recordr}
\title{Initialize a Recorder object}
\usage{
\S4method{initialize}{Recordr}(.Object, recordrDir = RecordrHome)
}
\arguments{
\item{.Object}{The Recordr object}

\item{recordrDir}{The directory to store provenance data in.}
}
\description{
Initialize a Recorder object
}
\seealso{
\code{\link[=Recordr-class]{Recordr}} { class description}
}

