% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recordr.R
\name{makeEML}
\alias{makeEML}
\title{Create a minimal EML document.
Creating EML should be more complete, but this minimal example will suffice to create a valid document.}
\usage{
makeEML(recordr, id, system, title, creators, abstract = NA,
  methodDescription = NA, geo_coverage = NA, temp_coverage = NA,
  endpoint = NA)
}
\description{
Create a minimal EML document.
Creating EML should be more complete, but this minimal example will suffice to create a valid document.
}

