% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvRels.R
\docType{class}
\name{ProvRels-class}
\alias{ProvRels-class}
\title{A class containing information about a file or group of files}
\description{
A class containing information about a file or group of files
}
\details{
This class is used internally by the recordr package.
}
\section{Slots}{

\describe{
\item{\code{executionId}}{a \code{characgter} containing the identifier associated with the file entry}

\item{\code{subject}}{a \code{character} containing the subject of a provenance relationship}

\item{\code{predicate}}{a \code{character} containign the predicate of a provenance relationship}

\item{\code{object}}{a \code{character} containing the object of a provenance relationship}

\item{\code{subjectType,}}{a \code{character} containing the RDF node type of the the subject, values can be 'uri', 'blank'}

\item{\code{objectType}}{a \code{character} containign the RDF node type of the object, each value can be 'uri', 'blank', or 'literal'}

\item{\code{dataTypeURI}}{The RDF data type that specifies the type of the object}
}}
\section{Methods}{

\itemize{
 \item{\code{\link[=initialize-ProvRels]{initialize}}}{: Initialize a ProvRels object}
 \item{\code{\link{readProvRels}}}{: Retrieve saved provenance relationships.}
 \item{\code{\link{writeProvRels}}}{: Save a provenance relationship.}object
}
}
\seealso{
\code{\link{recordr}}{ package description.}
}

