% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Recordr.R
\docType{methods}
\name{putMetadata}
\alias{putMetadata}
\alias{putMetadata,Recordr-method}
\title{Update the metadata object for a run}
\usage{
putMetadata(recordr, ...)

\S4method{putMetadata}{Recordr}(recordr, id = as.character(NA),
  seq = as.character(NA), metadata = as.character(NA), asText = TRUE)
}
\arguments{
\item{recordr}{a Recordr instance}

\item{id}{The identifier for a run}

\item{seq}{The sequence number for a run}

\item{metadata}{The replacement metadata, as the actual text, or as a filename containing the metadata}

\item{asText}{A logical. See 'Details'.
 If TRUE, then the \code{metadata} parameter is a character vector containing metadata, ir FALSE it is a filename. The
default is TRUE.}
}
\value{
A character vector containing the metadata
}
\description{
Put a metadata document into the recordr cache for an run, replacing the
existing metadata object for the specified run, if one exists.
}
\details{
The \code{metadata} parameter can specify either a character vector that contains the metadata
this parameter can be a filename that contains the metadata. The \code{asText} parameter is used to
specify which type of value is specified. If \code{asText} is TRUE, then the \code{metadata} parameter
is a character vector, if it is FALSE, then the \code{metadata} parameter is a filename.
}
\section{Methods (by class)}{
\itemize{
\item \code{Recordr}: 
}}
\seealso{
\code{\link[=Recordr-class]{Recordr}} { class description}
}

