% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvRels.R
\docType{methods}
\name{readProvRels}
\alias{readProvRels}
\alias{readProvRels,Recordr-method}
\title{Retrieve saved file metadata for one or more files}
\usage{
readProvRels(recordr, ...)

\S4method{readProvRels}{Recordr}(recordr, executionId = as.character(NA),
  subject = as.character(NA), predicate = as.character(NA),
  object = as.character(NA), subjectType = as.character(NA),
  objectType = as.character(NA), dataTypeURI = as.character(NA),
  orderBy = as.character(NA), sortOrder = "ascending", delete = FALSE,
  ...)
}
\arguments{
\item{recordr}{A recordr object}

\item{...}{Additional parameters}

\item{executionId}{A character value that specifies an execution identifier to search for.}

\item{subject}{The subject of the provenance relationships to match}

\item{predicate}{The predicate of the provenance relationships to match}

\item{object}{The object of the provenance relationships to match}

\item{subjectType}{A character value containing the subject type of the relationship to match}

\item{objectType}{A character value containing the object type of the relationship to match}

\item{dataTypeURI}{A character value containing the data type of the relationship to match}

\item{orderBy}{The column to sort the result set by.}

\item{sortOrder}{The sort type. Values include ("ascending", "descending")}

\item{delete}{a \code{"logical"}, if TRUE, the selected file entries are deleted (default: FALSE).}
}
\value{
A dataframe containing file metadata objects
}
\description{
File metadata is retrived from the recordr database table \emph{filemeta}
based on search parameters.
}
\details{
This method is used internally by the recordr package.
}
\seealso{
\code{\link[=ProvRels-class]{ProvRels}} { class description}
}
