% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{record}
\alias{record}
\title{Record provenance for an R script execution and create a DataONE DataPackage that
contains this provenance information and all derived products create by the script}
\usage{
record(recordr, fileName, ...)
}
\arguments{
\item{Recordr}{object}

\item{The}{filename of the R script to run and collect provenance information for}
}
\value{
the identifier for the DataONE datapackge created by this run
}
\description{
Record provenance for an R script execution and create a DataONE DataPackage that
contains this provenance information and all derived products create by the script
}
\author{
slaughter
}

