% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{recordr_create}
\alias{recordr_create}
\title{Provenance wrapper function for dataone::createObject method}
\usage{
recordr_create(mnode, pid, file, sysmeta, ...)
}
\arguments{
\item{mnode}{The member node to craete the object on}

\item{pid}{A persistent identifier}

\item{file}{The file to upload}

\item{sysmeta}{A SysstemMetadata object associated with the uploaded object}

\item{...}{additional parameters}
}
\description{
Override the dataone::createOjbect method and record a provenance relationship
for the object created.
}
\note{
This function is not intended to be called directly by a user.
}

