% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{recordr_createObject}
\alias{recordr_createObject}
\title{Provenance wrapper function for dataone::createObject method}
\usage{
recordr_createObject()
}
\description{
Override the dataone::createOjbect method and record a provenance relationship
for the object created.
}
\note{
This function is not intended to be called directly by a user.
}
