% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{recordr_ggsave}
\alias{recordr_ggsave}
\title{Provenance wrapper for the ggplot2::ggsave function}
\usage{
recordr_ggsave()
}
\description{
Override the ggplot2::ggsave function and record a provenance relationship
for the file that was written.
}
\note{
This function is not intended to be called directly by a user.
}
