% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{recordr_raster}
\alias{recordr_raster}
\title{Provenance wrapper for the raster::raster function}
\usage{
recordr_raster()
}
\description{
Override the raster::raster function and record a provenance relationship
for the file read.
}
\note{
This function is not intended to be called directly by a user.
}
