% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{recordr_read_csv}
\alias{recordr_read_csv}
\title{Provenance wrapper for the R readr::read_csv function}
\usage{
recordr_read_csv()
}
\arguments{
\item{...}{function parameters}
}
\description{
Override the readr::read_csv function and record a provenance relationship
for the file that was read.
}
\note{
This function is not intended to be called directly by a user.
}
