% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{recordr_readLines}
\alias{recordr_readLines}
\title{Provenance wrapper for R base::readLines function}
\usage{
recordr_readLines(con, ...)
}
\arguments{
\item{con}{A connetion to read from}

\item{...}{additional parameters}
}
\description{
Override the base::readLines function and record a provenance relationship
for the file read.
}
\note{
This function is not intended to be called directly by a user.
}

