% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{recordr_readOGR}
\alias{recordr_readOGR}
\title{Provenance wrapper for the rgdal::readOGR function}
\usage{
recordr_readOGR()
}
\arguments{
\item{dsn}{data source name}

\item{layer}{layer name}

\item{...}{additional parameters}
}
\description{
Override the rgdal::readOGR function and record a provenance relationship
for the file read.
}
\note{
This function is not intended to be called directly by a user.
}
