% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{recordr_readPNG}
\alias{recordr_readPNG}
\title{Provenance wrapper for the pnd::read function}
\usage{
recordr_readPNG()
}
\description{
Override the png::read function and record a provenance relationship
for the file read.
}
\note{
This function is not intended to be called directly by a user.
}
