% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{recordr_scan}
\alias{recordr_scan}
\title{Provenance wrapper for the R base::scan function}
\usage{
recordr_scan()
}
\arguments{
\item{file}{The file to scan.}

\item{...}{additional parameters}
}
\description{
Override the base::scan function and record a provenance relationship
for the scanned file.
}
\note{
This function is not intended to be called directly by a user.
}
