% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{recordr_updateObject}
\alias{recordr_updateObject}
\title{Provenance wrapper function for dataone::updateObject}
\usage{
recordr_updateObject()
}
\description{
Override the dataone::updateObject method and record a provenance
relationship for the object uploaded.
}
\note{
This function is not intended to be called directly by a user.
}
