% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{recordr_write_csv}
\alias{recordr_write_csv}
\title{Provenance wrapper for the R readr::write_csv function}
\usage{
recordr_write_csv()
}
\description{
Override the readr::write_csv function and record a provenance relationship
for the written file.
}
\note{
This function is not intended to be called directly by a user.
}
