% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{setProvCapture}
\alias{setProvCapture}
\title{Disable or enable provenance capture temporarily
It may be necessary to disable provenance capture temporarily, for example when
record() is writting out a housekeeping file.
A state variable in the ".recordr" environment is used to
temporarily disable provenance capture so that housekeeping tasks
will not have provenance information recorded for them.
Return the state of provenance capture: TRUE is enalbed, FALSE is disabled}
\usage{
setProvCapture(enable)
}
\arguments{
\item{enable}{logical variable used to enable or disable provenance capture}
}
\value{
enabled a logical indicating the state of provenance capture: TRUE=enabled, FALSE=disabled
}
\description{
Disable or enable provenance capture temporarily
It may be necessary to disable provenance capture temporarily, for example when
record() is writting out a housekeeping file.
A state variable in the ".recordr" environment is used to
temporarily disable provenance capture so that housekeeping tasks
will not have provenance information recorded for them.
Return the state of provenance capture: TRUE is enalbed, FALSE is disabled
}
\author{
slaughter
}

