% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{standardizeCall}
\alias{standardizeCall}
\title{Standardise a function call}
\usage{
standardizeCall(call, env = parent.frame())
}
\arguments{
\item{call}{A call}

\item{env}{Environment in which to look up call value.}
}
\description{
Standardise a function call
}
\note{
from Hadley Wicham's pryr standarize_call
}
