% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recordr.R
\name{upgradeRecordr}
\alias{upgradeRecordr}
\title{Update the recordr database to the current version}
\usage{
upgradeRecordr(recordr)
}
\arguments{
\item{recordr}{A recordr object}
}
\value{
logical TRUE if the upgrade was successful, FALSE if a problem was encountered.
}
\description{
Update the recordr database to the current version
}
