% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FileMetadata.R
\docType{methods}
\name{writeFileMeta}
\alias{writeFileMeta}
\alias{writeFileMeta,Recordr,FileMetadata-method}
\title{Save metadata for a single file.}
\usage{
writeFileMeta(recordr, fileMeta, ...)

\S4method{writeFileMeta}{Recordr,FileMetadata}(recordr, fileMeta, ...)
}
\arguments{
\item{recordr}{A recordr object}

\item{fileMeta}{A fileMetadata object}

\item{...}{(Not yet used)}
}
\description{
Metadata for a file is written to an RSQLite database.
}
\details{
This method is used internally by the recordr package.
}
\section{Methods (by class)}{
\itemize{
\item \code{recordr = Recordr,fileMeta = FileMetadata}: 
}}
\seealso{
\code{\link[=FileMetadata-class]{FileMetadata}} { class description}
}

