% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvRels.R
\docType{methods}
\name{writeProvRel}
\alias{writeProvRel}
\alias{writeProvRel,Recordr-method}
\title{Save a single provenance relationship.}
\usage{
writeProvRel(recordr, provRels, ...)

\S4method{writeProvRel}{Recordr}(recordr, provRels, ...)
}
\arguments{
\item{recordr}{A recordr object}

\item{provRels}{A ProvRels object.}

\item{...}{(Not yet used)}
}
\description{
Metadata for a provenance relationship is written to the recordr RSQLite database.
}
\details{
This method is used internally by the recordr package.
}
\seealso{
\code{\link[=ProvRels-class]{ProvRels}} { class description}
}
