% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ExecMetadata.R
\docType{class}
\name{ExecMetadata-class}
\alias{ExecMetadata-class}
\title{A class representing a script execution with the run manager}
\description{
A class representing a script execution with the run manager
}
\section{Slots}{

\describe{
\item{\code{executionId}}{}

\item{\code{metadataId}}{}

\item{\code{tag}}{}

\item{\code{datapackageId}}{}

\item{\code{user}}{}

\item{\code{subject}}{}

\item{\code{hostId}}{}

\item{\code{startTime}}{}

\item{\code{operatingSystem}}{}

\item{\code{runtime}}{}

\item{\code{softwareApplication}}{}

\item{\code{moduleDependencies}}{}

\item{\code{endTime}}{}

\item{\code{errorMessage}}{}

\item{\code{publishTime}}{}

\item{\code{publishNodeId}}{}

\item{\code{publishId}}{}

\item{\code{console}}{}

\item{\code{seq}}{}
}}
\section{Methods}{

\itemize{
 \item{\code{\link[=initialize-ExecMetadata]{initialize}}}{: Initialize an execution metadata object}
 \item{\code{\link{readExecMeta}}}{: Retrieve saved Execution metadata.}
 \item{\code{\link{writeExecMeta}}}{: Save a single execution metadat.}
 \item{\code{\link{updateExecMeta}}}{: Update saved execution metadata.}
}
}
\author{
slaughter
}
\seealso{
\code{\link{recordr}}{ package description.}
}

