% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FileMetadata.R
\docType{class}
\name{FileMetadata-class}
\alias{FileMetadata-class}
\title{A class containing information about a file or group of files}
\description{
A class containing information about a file or group of files
}
\details{
This class is used internally by the recordr package.
}
\section{Slots}{

\describe{
\item{\code{fileId}}{}

\item{\code{executionId}}{}

\item{\code{filePath}}{}

\item{\code{sha256}}{}

\item{\code{size}}{}

\item{\code{user}}{}

\item{\code{createTime}}{}

\item{\code{modifyTime}}{}

\item{\code{access}}{}

\item{\code{archivedFilePath}}{}
}}
\section{Methods}{

\itemize{
 \item{\code{\link[=initialize-FileMetadata]{initialize}}}{: Initialize a FileMetadata object}
 \item{\code{\link{readFileMeta}}}{: Retrieve saved file metadata for one or more files}
 \item{\code{\link{writeFileMeta}}}{: Save metadata for a single file.}
}
}
\seealso{
\code{\link{recordr}}{ package description.}
}

