% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ExecMetadata.R
\docType{methods}
\name{initialize,ExecMetadata-method}
\alias{initialize,ExecMetadata-method}
\alias{initialize-ExecMetadata}
\title{Initialize an execution metadata object}
\usage{
\S4method{initialize}{ExecMetadata}(.Object, programName = as.character(NA),
  tag = as.character(NA))
}
\arguments{
\item{.Object}{The ExecMetada object}

\item{programName}{The name of the program that is being run.}

\item{tag}{A character string that describes this execution.}
}
\description{
Initialize an execution metadata object
}
\seealso{
\code{\link[=ExecMetadata-class]{ExecMetadata}} { class description}
}

