% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Recordr.R
\docType{methods}
\name{publishRun}
\alias{publishRun}
\alias{publishRun,Recordr-method}
\title{Publish a recordr'd execution to DataONE}
\usage{
publishRun(recordr, id, ...)

\S4method{publishRun}{Recordr}(recordr, id, assignDOI = FALSE,
  update = FALSE)
}
\arguments{
\item{recordr}{a Recordr instance}

\item{id}{the run identifier for the package to upload to DataONE}

\item{assignDOI}{a boolean value: if TRUE, assign DOI values for system metadata, otherwise assign uuid values}

\item{update}{a boolean value: if TRUE, republish a previously published execution}
}
\value{
The published identifier of the uploaded package
}
\description{
Publish a recordr'd execution to DataONE
}
\section{Methods (by class)}{
\itemize{
\item \code{Recordr}: 
}}
\seealso{
\code{\link[=Recordr-class]{Recordr}} { class description}
}

