% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recordr.R
\docType{methods}
\name{publishRun}
\alias{publishRun}
\alias{publishRun,Recordr-method}
\title{Publish a recordr'd execution to DataONE}
\usage{
publishRun(recordr, ...)

\S4method{publishRun}{Recordr}(recordr, id = as.character(NA),
  seq = as.character(NA), assignDOI = FALSE, update = FALSE,
  quiet = TRUE)
}
\arguments{
\item{recordr}{a Recordr instance}

\item{...}{additional parameters
seealso \code{\link[=Recordr-class]{Recordr}} { class description}}

\item{id}{the run identifier for the execution to upload to DataONE}

\item{seq}{The sequence number for the execution to upload to DataONE}

\item{assignDOI}{a boolean value: if TRUE, assign DOI values for system metadata, otherwise assign uuid values}

\item{update}{a boolean value: if TRUE, republish a previously published execution}

\item{quiet}{A boolean value: if TRUE, informational messages are not printed (default=TRUE)}
}
\value{
The published identifier of the uploaded package
}
\description{
Publish a recordr'd execution to DataONE
}

