% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{recordr_write.csv}
\alias{recordr_write.csv}
\title{Override the R 'write.csv' method}
\usage{
recordr_write.csv(x, file, ...)
}
\description{
Record the provenance relationship of local objecct <- wasGeneratedBy <- script
}

