% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{recordr_write.csv}
\alias{recordr_write.csv}
\title{Provenance wrapper for the R write.csv function}
\usage{
recordr_write.csv()
}
\arguments{
\item{x}{The object to write}

\item{file}{The output connection to write to}

\item{...}{additional parameters}
}
\description{
Override the utils::write.csv function and record a provenance relationship
for the written file.
}
\note{
This function is not intended to be called directly by a user.
}
