% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{recordr_writeLines}
\alias{recordr_writeLines}
\title{Provenance wrapper for R base::writeLines function}
\usage{
recordr_writeLines(text, con, ...)
}
\arguments{
\item{text}{A character vector to write.}

\item{con}{The connection to write to.}

\item{...}{additional parameters}
}
\description{
Override the base::writeLines function and record a provenance relationship
for the file that was written.
}
\note{
This function is not intended to be called directly by a user.
}

