% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{recordr_writeOGR}
\alias{recordr_writeOGR}
\title{Provenance wrapper for the rgdal::writeOGR function}
\usage{
recordr_writeOGR()
}
\arguments{
\item{obj}{The image to write out.}

\item{dsn}{The data source name.}

\item{...}{additional parameters}
}
\description{
Override the rgdal::writeOGR function and record a provenance relationship
for the file that was written.
}
\note{
This function is not intended to be called directly by a user.
}
