% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{recordr_writePNG}
\alias{recordr_writePNG}
\title{Provenance wrapper for the png::write function}
\usage{
recordr_writePNG()
}
\description{
Override the png::write function and record a provenance relationship
for the file that was written.
}
\note{
This function is not intended to be called directly by a user.
}
