% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedFunctions.R
\name{recordr_writeRaster}
\alias{recordr_writeRaster}
\title{Provenance wrapper for the raster::writeRaster function}
\usage{
recordr_writeRaster()
}
\arguments{
\item{x}{Raster* object}

\item{file}{Output filename}

\item{format}{Output file type.}

\item{...}{additional parameters}
}
\value{
The name of the output file
}
\description{
Override the raster::writeRaster function and record a provenance relationship
for the file that was written.
}
\note{
This function is not intended to be called directly by a user.
}
