% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecMetadata.R
\docType{methods}
\name{writeExecMeta}
\alias{writeExecMeta}
\alias{writeExecMeta,Recordr,ExecMetadata-method}
\title{Save a single execution metadata.}
\usage{
writeExecMeta(recordr, execMeta, ...)

\S4method{writeExecMeta}{Recordr,ExecMetadata}(recordr, execMeta, ...)
}
\arguments{
\item{recordr}{A Recordr object}

\item{...}{Not yet used.}

\item{execMetadata}{an ExecMetadata object to save.}
}
\description{
Save a single execution metadata.
}
\seealso{
\code{\link[=ExecMetadata-class]{ExecMetadata}} { class description}
}

