% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvRels.R
\docType{methods}
\name{writeProvRel}
\alias{writeProvRel}
\alias{writeProvRel,Recordr-method}
\title{Save a single provenance relationship.}
\usage{
writeProvRel(recordr, provRels, ...)

\S4method{writeProvRel}{Recordr}(recordr, provRels, ...)
}
\arguments{
\item{recordr}{A recordr object}

\item{...}{(Not yet used)}

\item{subject}{The subject of a provenance relationship}

\item{predicate}{The predicate of a provenance relationship}

\item{object}{The object of a provenance relationship}

\item{subjectType}{a \code{"character"}, the RDF node type of the subject, i.e. "url", "blank"}

\item{objectType}{a \code{"character"}, the RDF node type of the object i.e. "url", "blank"}

\item{dataTypeURI}{a \code{"character"}, the RDF data type of the object, i.e. "xsd:string"}
}
\description{
Metadata for a file is written to an RSQLite database.
}
\details{
This method is used internally by the recordr package.
}
\seealso{
\code{\link[=ProvRels-class]{ProvRels}} { class description}
}

